backend/requirements.txt
httpx
anthropic
python-dotenv
